+++ Helena.Test.Back.Model/Entitys/CompanyEntity.cs
@@ -2,11 +2,11 @@
 {
     public class CompanyEntity
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string AvatarUrl { get; set; }
         public string NomeFantasia { get; set; }
         public string RazaoSocial { get; set; }
-        public int QtdeFuncionarios { get; set; }
+        public long QtdeFuncionarios { get; set; }
         public bool Active { get; set; }
     }
